修复从干簧管切换到震动开关时，休眠无法解除的问题 修复屏保计时器的一些隐藏问题 修复分钟转秒的时候发生精度损失的问题

+++ MathFun.h
@@ -6,7 +6,7 @@ int Get_Dec_Deep(int x);
 int Get_Int_Bit_Sum(int x);
 uint32_t Get_UTF8_Ascii_Pix_Len(uint8_t size, char* s);
 uint32_t UTF8_HMiddle(uint32_t x, uint32_t w, uint8_t size, char* s);
-uint32_t Minute2Millis(uint32_t min);
+uint32_t Minute2Millis(float min);
 
 void ShowBuffer(uint8_t* buffer, uint32_t size);
